Claude/review documentation 011 cv5oq a1rs pw ni s dghyr2b

+++ src/modules/Playerbot/Core/DI/Interfaces/IFarmingCoordinator.h
@@ -20,7 +20,7 @@ namespace Playerbot
 
 // Forward declarations
 enum class FarmingSessionType : uint8;
-enum class ProfessionType : uint8;
+enum class ProfessionType : uint16;  // Must match ProfessionManager.h (supports SkillType values)
 struct FarmingZoneInfo;
 struct FarmingSession;
 struct FarmingCoordinatorProfile;
